+++ Plugins/DbAndroid/dbandroidinstance.h
@@ -21,7 +21,8 @@ class DbAndroidInstance : public AbstractDb
 
         QList<AliasedColumn> columnsForQuery(const QString& query);
         SqlQueryPtr prepare(const QString& query);
-        QString getTypeLabel();
+        QString getTypeLabel() const;
+        QString getTypeClassName() const;
         bool deregisterFunction(const QString& name, int argCount);
         bool registerScalarFunction(const QString& name, int argCount, bool deterministic);
         bool registerAggregateFunction(const QString& name, int argCount, bool deterministic);