+++ SQLiteStudio3/coreSQLiteStudio/db/abstractdb3.h
@@ -59,7 +59,7 @@ class AbstractDb3 : public AbstractDb
         void initAfterOpen();
         SqlQueryPtr prepare(const QString& query);
         bool flushWalInternal();
-        QString getTypeLabel();
+        QString getTypeLabel() const;
         bool deregisterFunction(const QString& name, int argCount);
         bool registerScalarFunction(const QString& name, int argCount, bool deterministic);
         bool registerAggregateFunction(const QString& name, int argCount, bool deterministic);
@@ -487,7 +487,7 @@ SqlQueryPtr AbstractDb3<T>::prepare(const QString& query)
 }
 
 template <class T>
-QString AbstractDb3<T>::getTypeLabel()
+QString AbstractDb3<T>::getTypeLabel() const
 {
     return T::label;
 }