+++ SQLiteStudio3/coreSQLiteStudio/db/db.h
@@ -602,7 +602,16 @@ class API_EXPORT Db : public QObject, public Interruptable
          * This is usually the same as DbPlugin::getTitle(), but getTitle() is used in list of plugins in configuration dialog,
          * while getTypeLabel() is used on databases list.
          */
-        virtual QString getTypeLabel() = 0;
+        virtual QString getTypeLabel() const = 0;
+
+        /**
+         * @brief Gets C++ class name implementing this particular Db instance.
+         * @return Class name.
+         *
+         * It can be used to distinguish between different drivers of Db instances. While getTypeLabel() can theoretically return
+         * same labels for two different drivers, this method will always return distinct class name.
+         */
+        virtual QString getTypeClassName() const = 0;
 
         /**
          * @brief Initializes resources once the all derived Db classes are constructed.