+++ SQLiteStudio3/coreSQLiteStudio/db/dbsqlite3.cpp
@@ -19,3 +19,8 @@ Db* DbSqlite3::clone() const
 {
     return new DbSqlite3(name, path, connOptions);
 }
+
+QString DbSqlite3::getTypeClassName() const
+{
+    return "DbSqlite3";
+}