+++ SQLiteStudio3/coreSQLiteStudio/db/invaliddb.h
@@ -48,7 +48,8 @@ class API_EXPORT InvalidDb : public Db
         QString getUniqueNewObjectName(const QString& attachedDbName);
         QString getErrorText();
         int getErrorCode();
-        QString getTypeLabel();
+        QString getTypeLabel() const;
+        QString getTypeClassName() const;
         bool initAfterCreated();
         bool deregisterFunction(const QString& name, int argCount);
         bool registerScalarFunction(const QString& name, int argCount, bool deterministic);