+++ SQLiteStudio3/coreSQLiteStudio/dbobjectorganizer.cpp
@@ -198,7 +198,7 @@ bool DbObjectOrganizer::processAll()
 
     // Attaching target db if needed
     AttachGuard attach;
-    if (!(referencedTables + srcTables).isEmpty())
+    if (srcDb->getTypeClassName() == dstDb->getTypeClassName() && !(referencedTables + srcTables).isEmpty())
     {
         attach = srcDb->guardedAttach(dstDb, true);
         attachName = attach->getName();
@@ -431,12 +431,16 @@ bool DbObjectOrganizer::copyTableToDb(const QString& table)
 
 bool DbObjectOrganizer::copyDataAsMiddleware(const QString& table)
 {
-    QStringList srcColumns = srcResolver->getTableColumns(srcTable);
+    static_qstring(selectTpl, "SELECT %1 FROM %2");
+    static_qstring(insertTpl, "INSERT INTO %1 (%2) VALUES (%3)");
+
+    QStringList srcColumns = srcResolver->getTableColumns(srcTable, true);
+    QString srcColumnsStr = srcColumns.join(", ");
     QString wrappedSrcTable = wrapObjIfNeeded(srcTable);
-    SqlQueryPtr results = srcDb->prepare("SELECT * FROM " + wrappedSrcTable);
+    SqlQueryPtr results = srcDb->prepare(selectTpl.arg(srcColumnsStr, wrappedSrcTable));
     if (!results->execute())
     {
-        notifyError(tr("Error while copying data for table %1: %2").arg(table).arg(results->getErrorText()));
+        notifyError(tr("Error while copying data for table %1: %2").arg(table, results->getErrorText()));
         return false;
     }
 
@@ -445,7 +449,7 @@ bool DbObjectOrganizer::copyDataAsMiddleware(const QString& table)
         argPlaceholderList << "?";
 
     QString wrappedDstTable = wrapObjIfNeeded(table);
-    QString sql = "INSERT INTO " + wrappedDstTable + " VALUES (" + argPlaceholderList.join(", ") + ")";
+    QString sql = insertTpl.arg(wrappedDstTable, srcColumnsStr, argPlaceholderList.join(", "));
     SqlQueryPtr insertQuery = dstDb->prepare(sql);
 
     SqlResultsRowPtr row;
@@ -455,14 +459,14 @@ bool DbObjectOrganizer::copyDataAsMiddleware(const QString& table)
         row = results->next();
         if (!row)
         {
-            notifyError(tr("Error while copying data to table %1: %2").arg(table).arg(results->getErrorText()));
+            notifyError(tr("Error while copying data to table %1: %2").arg(table, results->getErrorText()));
             return false;
         }
 
         insertQuery->setArgs(row->valueList());
         if (!insertQuery->execute())
         {
-            notifyError(tr("Error while copying data to table %1: %2").arg(table).arg(insertQuery->getErrorText()));
+            notifyError(tr("Error while copying data to table %1: %2").arg(table, insertQuery->getErrorText()));
             return false;
         }
 
@@ -480,12 +484,16 @@ bool DbObjectOrganizer::copyDataAsMiddleware(const QString& table)
 
 bool DbObjectOrganizer::copyDataUsingAttach(const QString& table)
 {
+    static_qstring(insertTpl, "INSERT INTO %1.%2 (%3) SELECT %3 FROM %4");
+
     QString wrappedSrcTable = wrapObjIfNeeded(srcTable);
     QString wrappedDstTable = wrapObjIfNeeded(table);
-    SqlQueryPtr results = srcDb->exec("INSERT INTO " + attachName + "." + wrappedDstTable + " SELECT * FROM " + wrappedSrcTable);
+    QStringList srcColumns = srcResolver->getTableColumns(srcTable, true);
+    QString srcColumnsStr = srcColumns.join(", ");
+    SqlQueryPtr results = srcDb->exec(insertTpl.arg(attachName, wrappedDstTable, srcColumnsStr, wrappedSrcTable));
     if (results->isError())
     {
-        notifyError(tr("Error while copying data to table %1: %2").arg(table).arg(results->getErrorText()));
+        notifyError(tr("Error while copying data to table %1: %2").arg(table, results->getErrorText()));
         return false;
     }
     return true;
@@ -508,7 +516,7 @@ void DbObjectOrganizer::dropObject(const QString& name, const QString& type)
     if (results->isError())
     {
         notifyWarn(tr("Error while dropping source view %1: %2\nTables, indexes, triggers and views copied to database %3 will remain.")
-                   .arg(name).arg(results->getErrorText()).arg(dstDb->getName()));
+                   .arg(name, results->getErrorText(), dstDb->getName()));
     }
 }