+++ ChangeLog.md
@@ -3,6 +3,7 @@
 ### 3.4.2
 - BUGFIX: #4602 Fixed editing data in table with Russian upper-case name.
 - BUGFIX: #4105 Fixed Import Dialog state update when changing import source (CSV vs RegExp), which could cause inability to proceed with the import in some cases.
+- BUGFIX: #3767 Fixed copying objects between unencrypted and encrypted databases, also fixed copying tables with generated columns.
 - BUGFIX: Fixed refreshing invalid database state if edited & fixed its connection options.
 
 ### 3.4.1